created dashboard to see data

+++ caching/caching.py
@@ -1,16 +1,21 @@
 import yfinance as yf
 import json
 import os
+import logging
+
+logger = logging.getLogger(__name__)
 
 def cache_market_caps(tickers: list[str]):
     mcaps = dict()
     for ticker in tickers:
         try:
             mcaps[ticker] = yf.Ticker(ticker).info["marketCap"]
-        except:
-            print(ticker)
+        except KeyError:
+            logger.warning(f"No market cap data for {ticker}")
 
-    os.remove("./caching/market_caps.json")
+    if os.path.exists("./caching/market_caps.json"):
+        os.remove("./caching/market_caps.json")
+
     with open("./caching/market_caps.json", "w") as f:
         json.dump(mcaps, f)
